Custom item prosthetics fix (#20462)
This adds a toggle to the equip_custom_items proc to alternate between
it equipping only body-related items, and everything else. These two
toggleable modes are called separately.

This is intended to resolve a bug in live where, if you join in with a
custom item prosthetic leg, your shoes will be automatically unequipped
the moment you join. This seems to happen because custom items are
currently equipped after the rest of the loadout/job equipment, so your
shoes are put on before your custom item leg is.

+++ code/modules/customitems/item_spawning.dm
@@ -216,7 +216,9 @@ GLOBAL_LIST_INIT(custom_items, list())
 GLOBAL_LIST_EMPTY(character_id_to_custom_items_mapping)
 
 /// Gets the relevant custom items for the given target_mob, character_id, and player_ckey, and equips the target_mob with those custom items
-/proc/equip_custom_items(var/mob/living/carbon/human/target_mob, var/character_id, var/player_ckey)
+/// If body_only is true, it will only equip augments and prosthetics. If body_only is false, it will equip everything else.
+/// You will need to call this twice, once true, once false. This distinction exists to avoid awkward behaviour with prosthetic custom items.
+/proc/equip_custom_items(var/mob/living/carbon/human/target_mob, var/character_id, var/player_ckey, var/body_only = FALSE)
 	if(!character_id)
 		character_id = target_mob.character_id
 
@@ -267,10 +269,13 @@ GLOBAL_LIST_EMPTY(character_id_to_custom_items_mapping)
 			GLOB.character_id_to_custom_items_mapping[glob_character_id_key] = custom_items_list
 
 	for(var/datum/custom_item/ci as anything in GLOB.character_id_to_custom_items_mapping[glob_character_id_key])
-		equip_custom_item_to_mob(ci, target_mob)
+		equip_custom_item_to_mob(ci, target_mob, body_only)
 
-
-/proc/equip_custom_item_to_mob(var/datum/custom_item/citem, var/mob/living/carbon/human/target_mob)
+/// This equips custom items to mobs. The body_only parameter determines if it will equip non-body related augments (i.e. prosthetics or augments).
+/// If set to true, it only equips prosthetics and augments. If set to false, it only equips other custom items, such as clothes or accessories.
+/// It should be called twice, true then false - augments and prosthetics should be added before any clothing items, loadout or custom.
+/// This is so you don't get shoes falling off because someone's custom prosthetic foot was loaded in after their footwear, for instance.
+/proc/equip_custom_item_to_mob(var/datum/custom_item/citem, var/mob/living/carbon/human/target_mob, var/body_only = FALSE)
 	// Check for required job title.
 	if(length(citem.req_titles))
 		var/has_title
@@ -283,7 +288,8 @@ GLOBAL_LIST_EMPTY(character_id_to_custom_items_mapping)
 			to_chat(target_mob, "A custom item could not be equipped as you have joined with the wrong role.")
 			return FALSE
 
-	if(ispath(citem.item_path, /obj/item/organ/internal/augment/fluff))
+	// Next two conditionals only proc if body_only is true, otherwise they do not. If body_only is true, only these proc.
+	if(body_only && ispath(citem.item_path, /obj/item/organ/internal/augment/fluff))
 		var/obj/item/organ/internal/augment/fluff/aug = citem.spawn_item(target_mob)
 		var/obj/item/organ/external/affected = target_mob.get_organ(aug.parent_organ)
 		aug.replaced(target_mob, affected)
@@ -292,21 +298,23 @@ GLOBAL_LIST_EMPTY(character_id_to_custom_items_mapping)
 		target_mob.UpdateDamageIcon()
 		return
 
-	if(ispath(citem.item_path, /obj/item/organ/external))
-		var/obj/item/organ/external/dummy_limb = citem.item_path
-		var/obj/item/organ/external/parent_organ = target_mob.get_organ(initial(dummy_limb.parent_organ))
-		if(parent_organ)
-			var/obj/item/organ/external/existing_limb = target_mob.get_organ(initial(dummy_limb.limb_name))
-			if(existing_limb)
-				existing_limb.droplimb(TRUE)
-				qdel(existing_limb)
+	if(body_only)
+		if(ispath(citem.item_path, /obj/item/organ/external))
+			var/obj/item/organ/external/dummy_limb = citem.item_path
+			var/obj/item/organ/external/parent_organ = target_mob.get_organ(initial(dummy_limb.parent_organ))
+			if(parent_organ)
+				var/obj/item/organ/external/existing_limb = target_mob.get_organ(initial(dummy_limb.limb_name))
+				if(existing_limb)
+					existing_limb.droplimb(TRUE)
+					qdel(existing_limb)
 
-			var/obj/item/organ/external/custom_limb = citem.spawn_item(target_mob)
-			custom_limb.replaced(target_mob, parent_organ)
+				var/obj/item/organ/external/custom_limb = citem.spawn_item(target_mob)
+				custom_limb.replaced(target_mob, parent_organ)
 
-			target_mob.update_body()
-			target_mob.updatehealth()
-			target_mob.UpdateDamageIcon()
+				target_mob.update_body()
+				target_mob.updatehealth()
+				target_mob.UpdateDamageIcon()
+		// This return cuts the proc early if it should only be doing body-related items, i.e. augments and prosthetics.
 		return
 
 	// ID cards and MCs are applied directly to the existing object rather than spawned fresh.
